🎨 Move from_values from Record to CanValidate
